refactor: use workspace for module separation
The whole monotax crate is now called monotax-core. I'll split it into
fine-grained modules later.

fix(test): fix doctests imports

refactor: use underscore in crate names

refactor: add updated lock file
