+++ Cargo.toml
@@ -1,34 +1,3 @@
-[package]
-name = "monotax"
-version = "0.3.0-alpha.2"
-edition = "2021"
-description = "A tool to generate and manage tax reports"
-license-file = "LICENSE"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0.81"
-chrono = "0.4.37"
-clap = { version = "4.5.4", features = ["derive"] }
-csv = "1.3.0"
-directories = "5.0.1"
-encoding_rs = "0.8.33"
-encoding_rs_rw = "0.4.2"
-env_logger = "0.11.3"
-log = "0.4.21"
-serde = { version = "1.0.197", features = ["derive"] }
-tokio = { version = "1", features = ["rt", "rt-multi-thread", "macros"] }
-
-sqlx = { version = "0.7", features = [
-    "macros",
-    "sqlite",
-    "runtime-tokio",
-    "chrono",
-] }
-toml = "0.8.12"
-dotenvy = "0.15.7"
-uuid = { version = "1.9.1", features = ["v4"] }
-async-trait = "0.1.81"
-config = "0.14.1"
-
+[workspace]
+members = ["crates/*"]
+resolver = "2"
