+++ crates/monotax_core/src/domain/filter/income.rs
@@ -16,8 +16,8 @@
 //! All criteria are combined into a single [`IncomeCriteria`] object.
 //!
 //! ```rust
-//! use monotax::domain::filter::income::{IncomeCriteria, IncomeCriterion, YearFilter, QuarterFilter};
-//! use monotax::domain::Quarter;
+//! use monotax_core::domain::filter::income::{IncomeCriteria, IncomeCriterion, YearFilter, QuarterFilter};
+//! use monotax_core::domain::Quarter;
 //!
 //! let criteria = IncomeCriteria::new(&[
 //!     IncomeCriterion::Year(YearFilter::One(2021)),
