+++ crates/monotax_core/src/payment/report/plaintext.rs
