+++ crates/monotax_core/src/report/console.rs
