+++ crates/monotax_core/tests/income_repository_ctk.rs
@@ -1,11 +1,11 @@
 use chrono::NaiveDateTime;
-use monotax::domain::filter::income::IncomeCriteria;
-use monotax::domain::filter::income::IncomeCriterion;
-use monotax::domain::filter::income::QuarterFilter;
-use monotax::domain::filter::income::YearFilter;
-use monotax::domain::repository::IncomeRepository;
-use monotax::domain::Income;
-use monotax::domain::Quarter;
+use monotax_core::domain::filter::income::IncomeCriteria;
+use monotax_core::domain::filter::income::IncomeCriterion;
+use monotax_core::domain::filter::income::QuarterFilter;
+use monotax_core::domain::filter::income::YearFilter;
+use monotax_core::domain::repository::IncomeRepository;
+use monotax_core::domain::Income;
+use monotax_core::domain::Quarter;
 
 fn income(datetime: &str, amount: f64) -> Income {
     let datetime = NaiveDateTime::parse_from_str(datetime, "%Y-%m-%d %H:%M:%S").unwrap();
